#25 PDF export util implementation, Raw data moved to a service, HTML2Canvas removed from dependencies
