+++ package.json
@@ -48,7 +48,6 @@
     "@angular/router": "^9.1.7",
     "@angular/service-worker": "^9.1.7",
     "@nicky-lenaers/ngx-scroll-to": "^9.0.0",
-    "html2canvas": "^1.0.0-rc.5",
     "jspdf": "^1.5.3",
     "ngx-mask": "^9.0.2",
     "rxjs": "~6.5.4",
@@ -64,7 +63,6 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.12.39",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@types/html2canvas": "0.0.35",
     "@types/jspdf": "^1.3.3",
     "codelyzer": "^5.1.2",
     "eslint": "^7.0.0",
