+++ src/app/components/calculate/results/results.component.html
@@ -3,8 +3,7 @@
           [fxFlex.xl]="'0 1 50'"
           [fxFlex.lg]="'0 1 60'"
           [fxFlex.md]="'0 1 75'"
-          [fxFlex.lt-md]="'0 1 90'"
-          id="content">
+          [fxFlex.lt-md]="'0 1 90'">
   <mat-list fxFlex="0 1 85"
             [fxFlex.lt-md]="'0 1 100'"
             fxLayout="column"
